+++ src/app/(root)/(payment)/check-payment/_components/CheckPaymentContent.tsx
