+++ src/app/(root)/(payment)/payment/page.tsx
@@ -1,0 +1,5 @@
+const Payment = () => {
+  return <div className="mt-[160px]">Payment</div>;
+};
+
+export default Payment;
