+++ src/app/(root)/pay-history/page.tsx
@@ -1,6 +1,6 @@
 //주문 내역 페이지
 
-import PayHistoryList from '@/app/(root)/payment/_components/PayHistoryList';
+import PayHistoryList from '@/app/(root)/pay-history/_components/PayHistoryList';
 
 const PaymentList = () => {
   return (
